+++ src/Types/AnyType.php
@@ -15,6 +15,17 @@ class AnyType extends Type
     protected $regex = '(?<%s>.*)';
 
     /**
+     * Allows to write to a cached static configuration file
+     * 
+     * @param  array  $array cached variables
+     * @return object
+     */
+    public static function __set_state(array $array)
+    {
+        return new Self($array['pattern'], $array['regex']);
+    }
+
+    /**
      * Php format
      *
      * @param  number $value